# refactor: rewrite the check for unknown keys in the machine configuration