… pkg/machinery/config/decoder/testdata/controlplane.yaml
… pkg/machinery/config/decoder/testdata/worker.yaml