Fixes #4687 

This avoids doing multiple rounds of marshal/unmarshal to check for
unknown keys.

New version simply walks the type tree of the `*v1alpha1.Config` and
maps that to the values in YAML.

Benchmark result comparing old and new implementations:

```
~/go/bin/benchstat old.txt new.txt                  7.9s  Ср 15 дек 2021 21:32:00
name                      old time/op    new time/op    delta
DecoderV1Alpha1Config-16    7.83ms ± 7%    0.98ms ±22%  -87.50%  (p=0.008 n=5+5)

name                      old alloc/op   new alloc/op   delta
DecoderV1Alpha1Config-16    6.39MB ± 0%    0.24MB ± 0%  -96.24%  (p=0.008 n=5+5)

name                      old allocs/op  new allocs/op  delta
DecoderV1Alpha1Config-16     25.7k ± 0%      2.4k ± 0%  -90.55%  (p=0.008 n=5+5)
```

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/talos-systems/talos/4695)
<!-- Reviewable:end -->